Follow up on pull request #552

+++ src/FoldersTree.m
@@ -853,7 +853,7 @@
 /* numberOfChildrenOfItem
  * Returns the number of children belonging to the specified item
  */
--(int)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
+-(NSInteger)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
 {
 	TreeNode * node = (TreeNode *)item;
 	if (node == nil)
@@ -864,7 +864,7 @@
 /* child
  * Returns the child at the specified offset of the item
  */
--(id)outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item
+-(id)outlineView:(NSOutlineView *)outlineView child:(NSInteger)index ofItem:(id)item
 {
 	TreeNode * node = (TreeNode *)item;
 	if (node == nil)
@@ -1096,7 +1096,7 @@
  * Called when something is being dragged over us. We respond with an NSDragOperation value indicating the
  * feedback for the user given where we are.
  */
--(NSDragOperation)outlineView:(NSOutlineView*)olv validateDrop:(id <NSDraggingInfo>)info proposedItem:(id)item proposedChildIndex:(int)index
+-(NSDragOperation)outlineView:(NSOutlineView*)olv validateDrop:(id <NSDraggingInfo>)info proposedItem:(id)item proposedChildIndex:(NSInteger)index
 {
 	NSPasteboard * pb = [info draggingPasteboard]; 
 	NSString * type = [pb availableTypeFromArray:[NSArray arrayWithObjects:MA_PBoardType_FolderList, MA_PBoardType_RSSSource, @"WebURLsWithTitlesPboardType", NSStringPboardType, nil]]; 
@@ -1372,7 +1372,7 @@
 /* acceptDrop
  * Accept a drop on or between nodes either from within the folder view or from outside.
  */
--(BOOL)outlineView:(NSOutlineView *)olv acceptDrop:(id <NSDraggingInfo>)info item:(id)targetItem childIndex:(int)child
+-(BOOL)outlineView:(NSOutlineView *)olv acceptDrop:(id <NSDraggingInfo>)info item:(id)targetItem childIndex:(NSInteger)child
 { 
 	__block int childIndex = child;
 	NSPasteboard * pb = [info draggingPasteboard];
